mw-teardown: Only cp LocalSettings if it exists
If the job failed early during the installer, then this will not
yet exist. In that case, we want to make sure that we still
archive log/ files for inspection, and still clean up tmp dir.

+++ bin/mw-teardown.sh
@@ -4,4 +4,6 @@
 . /srv/deployment/integration/slave-scripts/bin/mw-set-env.sh
 
 # Copy LocalSettings to /log so that archive-log-dir collects it
-cp "${MW_INSTALL_PATH}/LocalSettings.php" "$LOG_DIR"
+if [ -e "${MW_INSTALL_PATH}/LocalSettings.php" ]; then
+	cp "${MW_INSTALL_PATH}/LocalSettings.php" "$LOG_DIR"
+fi
